+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Template for next version
 
 ### Fixed
 
+- error sending notification when field is not defined (https://github.com/ohmyform/ohmyform/issues/161)
+
 ### Security
 
 ## [1.0.1] - 2022-03-01